requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
streamlit==1.28.0
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
python-multipart==0.0.6
